refactor binary depedencies handling outside workspace class
motivation: code handling binary dependencies has become complex and large, factor it out to its own class

changes:
* factor out the code that deals with downloading and extracting binary dependencies out of Workspace and into a dedicated struct
* adjust workspace initiatilizers to allow customizing the BinaryArtifactsManager instead of the HTTPClient and Archiver that are in service of handling binary dependencies
* adjjust call sites and test
