+++ Sources/Workspace/Workspace.swift
@@ -186,6 +186,31 @@ private struct WorkspaceDependencyResolverDelegate: DependencyResolverDelegate {
     func failedToResolve(incompatibility: Incompatibility) {}
     func solved(result: [(package: PackageReference, binding: BoundVersion, products: ProductFilter)]) {}
 }
+
+private class WorkspaceBinaryArtifactsManagerDelegate: Workspace.BinaryArtifactsManager.Delegate {
+    private unowned let workspaceDelegate: Workspace.Delegate
+
+    init(workspaceDelegate: Workspace.Delegate) {
+        self.workspaceDelegate = workspaceDelegate
+    }
+
+    func willDownloadBinaryArtifact(from url: String) {
+        self.workspaceDelegate.willDownloadBinaryArtifact(from: url)
+    }
+
+    func didDownloadBinaryArtifact(from url: String, result: Result<AbsolutePath, Error>, duration: DispatchTimeInterval) {
+        self.workspaceDelegate.didDownloadBinaryArtifact(from: url, result: result, duration: duration)
+    }
+
+    func downloadingBinaryArtifact(from url: String, bytesDownloaded: Int64, totalBytesToDownload: Int64?) {
+        self.workspaceDelegate.downloadingBinaryArtifact(from: url, bytesDownloaded: bytesDownloaded, totalBytesToDownload: totalBytesToDownload)
+    }
+
+    func didDownloadAllBinaryArtifacts() {
+        self.workspaceDelegate.didDownloadAllBinaryArtifacts()
+    }
+}
+
 /// A workspace represents the state of a working project directory.
 ///
 /// The workspace is responsible for managing the persistent working state of a
@@ -245,7 +270,7 @@ public class Workspace {
         return self.customPackageContainerProvider ?? self
     }
 
-    /// The repository manager.
+    /// Source control repository manager used for interacting with source control based dependencies
     // var for backwards compatibility with deprecated initializers, remove with them
     fileprivate var repositoryManager: RepositoryManager
 
@@ -253,22 +278,16 @@ public class Workspace {
     // var for backwards compatibility with deprecated initializers, remove with them
     fileprivate var registryClient: RegistryClient
 
+    /// Registry based dependencies downloads manager used for interacting with registry based dependencies
     fileprivate var registryDownloadsManager: RegistryDownloadsManager
 
-    /// The http client used for downloading binary artifacts.
-    fileprivate let httpClient: HTTPClient
-
-    fileprivate let authorizationProvider: AuthorizationProvider?
-
-    /// The downloader used for unarchiving binary artifacts.
-    fileprivate let archiver: Archiver
-
-    /// The algorithm used for generating file checksums.
-    fileprivate let checksumAlgorithm: HashAlgorithm
+    /// Binary artifacts manager used for downloading and extracting binary artifacts
+    fileprivate let binaryArtifactsManager: BinaryArtifactsManager
     
     /// The package fingerprints storage
     fileprivate let fingerprints: PackageFingerprintStorage?
 
+    /// The workspace configuration settings
     fileprivate let configuration: WorkspaceConfiguration
 
     // state
@@ -328,9 +347,8 @@ public class Workspace {
             customRepositoryManager: .none,
             customRepositoryProvider: customRepositoryProvider,
             customRegistryClient: .none,
+            customBinaryArtifactsManager: .none,
             customIdentityResolver: .none,
-            customHTTPClient: .none,
-            customArchiver: .none,
             customChecksumAlgorithm: .none,
             delegate: delegate
         )
@@ -478,9 +496,11 @@ public class Workspace {
             customRepositoryManager: customRepositoryManager,
             customRepositoryProvider: customRepositoryProvider,
             customRegistryClient: customRegistryClient,
+            customBinaryArtifactsManager: .init(
+                httpClient: customHTTPClient,
+                archiver: customArchiver
+            ),
             customIdentityResolver: customIdentityResolver,
-            customHTTPClient: customHTTPClient,
-            customArchiver: customArchiver,
             customChecksumAlgorithm: customChecksumAlgorithm,
             delegate: delegate
         )
@@ -596,9 +616,8 @@ public class Workspace {
         customRepositoryManager: RepositoryManager? = .none,
         customRepositoryProvider: RepositoryProvider? = .none,
         customRegistryClient: RegistryClient? = .none,
+        customBinaryArtifactsManager: CustomBinaryArtifactsManager? = .none,
         customIdentityResolver: IdentityResolver? = .none,
-        customHTTPClient: HTTPClient? = .none,
-        customArchiver: Archiver? = .none,
         customChecksumAlgorithm: HashAlgorithm? = .none,
         // delegate
         delegate: Delegate? = .none
@@ -619,9 +638,8 @@ public class Workspace {
             customRepositoryManager: customRepositoryManager,
             customRepositoryProvider: customRepositoryProvider,
             customRegistryClient: customRegistryClient,
+            customBinaryArtifactsManager: customBinaryArtifactsManager,
             customIdentityResolver: customIdentityResolver,
-            customHTTPClient: customHTTPClient,
-            customArchiver: customArchiver,
             customChecksumAlgorithm: customChecksumAlgorithm,
             delegate: delegate
         )
@@ -645,9 +663,8 @@ public class Workspace {
         customRepositoryManager: RepositoryManager?,
         customRepositoryProvider: RepositoryProvider?,
         customRegistryClient: RegistryClient?,
+        customBinaryArtifactsManager: CustomBinaryArtifactsManager?,
         customIdentityResolver: IdentityResolver?,
-        customHTTPClient: HTTPClient?,
-        customArchiver: Archiver?,
         customChecksumAlgorithm: HashAlgorithm?,
         // delegate
         delegate: Delegate?
@@ -724,27 +741,35 @@ public class Workspace {
             )
         }
 
-        let httpClient = customHTTPClient ?? HTTPClient()
-        let archiver = customArchiver ?? ZipArchiver(fileSystem: fileSystem)
+        let binaryArtifactsManager = BinaryArtifactsManager(
+            fileSystem: fileSystem,
+            authorizationProvider: authorizationProvider,
+            hostToolchain: hostToolchain,
+            checksumAlgorithm: checksumAlgorithm,
+            customHTTPClient: customBinaryArtifactsManager?.httpClient,
+            customArchiver: customBinaryArtifactsManager?.archiver,
+            delegate: delegate.map(WorkspaceBinaryArtifactsManagerDelegate.init(workspaceDelegate:))
+        )
 
         // initialize
         self.fileSystem = fileSystem
+        self.configuration = configuration
         self.location = location
         self.delegate = delegate
         self.mirrors = mirrors
-        self.authorizationProvider = authorizationProvider
+
         self.hostToolchain = hostToolchain
         self.manifestLoader = manifestLoader
         self.currentToolsVersion = currentToolsVersion
         self.toolsVersionLoader = toolsVersionLoader
-        self.httpClient = httpClient
-        self.archiver = archiver
+
         self.customPackageContainerProvider = customPackageContainerProvider
         self.repositoryManager = repositoryManager
         self.registryClient = registryClient
         self.registryDownloadsManager = registryDownloadsManager
+        self.binaryArtifactsManager = binaryArtifactsManager
+
         self.identityResolver = identityResolver
-        self.checksumAlgorithm = checksumAlgorithm
         self.fingerprints = fingerprints
 
         self.pinsStore = LoadableResult {
@@ -755,8 +780,6 @@ public class Workspace {
                 mirrors: mirrors
             )
         }
-
-        self.configuration = configuration
 
         self.state = WorkspaceState(
             fileSystem: fileSystem,
@@ -1362,23 +1385,6 @@ extension Workspace {
             }
             completion(result)
         }
-    }
-
-    /// Generates the checksum
-    public func checksum(forBinaryArtifactAt path: AbsolutePath) throws -> String {
-        // Validate the path has a supported extension.
-        guard let pathExtension = path.extension, archiver.supportedExtensions.contains(pathExtension) else {
-            let supportedExtensionList = archiver.supportedExtensions.joined(separator: ", ")
-            throw StringError("unexpected file type; supported extensions are: \(supportedExtensionList)")
-        }
-
-        // Ensure that the path with the accepted extension is a file.
-        guard fileSystem.isFile(path) else {
-            throw StringError("file not found at path: \(path.pathString)")
-        }
-
-        let contents = try fileSystem.readFileContents(path)
-        return self.checksumAlgorithm.hash(contents).hexadecimalRepresentation
     }
 
     /// Returns `true` if the file at the given path might influence build settings for a `swiftc` or `clang` invocation generated by SwiftPM.
@@ -2298,11 +2304,11 @@ extension Workspace {
         addedOrUpdatedPackages: [PackageReference],
         observabilityScope: ObservabilityScope
     ) throws {
-        let manifestArtifacts = try self.parseArtifacts(from: manifests, observabilityScope: observabilityScope)
+        let manifestArtifacts = try self.binaryArtifactsManager.parseArtifacts(from: manifests, observabilityScope: observabilityScope)
 
         var artifactsToRemove: [ManagedArtifact] = []
         var artifactsToAdd: [ManagedArtifact] = []
-        var artifactsToDownload: [RemoteArtifact] = []
+        var artifactsToDownload: [BinaryArtifactsManager.RemoteArtifact] = []
         var artifactsToExtract: [ManagedArtifact] = []
 
         for artifact in state.artifacts {
@@ -2318,10 +2324,10 @@ extension Workspace {
                 targetName: artifact.targetName
             ]
 
-            if let fileExtension = artifact.path.extension, self.manifestLoader.supportedArchiveExtensions.contains(fileExtension) {
+            if artifact.path.extension?.lowercased() == "zip" {
                 // If we already have an artifact that was extracted from an archive with the same checksum,
                 // we don't need to extract the artifact again.
-                if case .local(let existingChecksum) = existingArtifact?.source, existingChecksum == (try self.checksum(forBinaryArtifactAt: artifact.path)) {
+                if case .local(let existingChecksum) = existingArtifact?.source, existingChecksum == (try self.binaryArtifactsManager.checksum(forBinaryArtifactAt: artifact.path)) {
                     continue
                 }
 
@@ -2392,11 +2398,19 @@ extension Workspace {
         }
 
         // Download the artifacts
-        let downloadedArtifacts = try self.download(artifactsToDownload, observabilityScope: observabilityScope)
+        let downloadedArtifacts = try self.binaryArtifactsManager.download(
+            artifactsToDownload,
+            artifactsDirectory: self.location.artifactsDirectory,
+            observabilityScope: observabilityScope
+        )
         artifactsToAdd.append(contentsOf: downloadedArtifacts)
 
         // Extract the local archived artifacts
-        let extractedLocalArtifacts = try self.extract(artifactsToExtract, observabilityScope: observabilityScope)
+        let extractedLocalArtifacts = try self.binaryArtifactsManager.extract(
+            artifactsToExtract,
+            artifactsDirectory: self.location.artifactsDirectory,
+            observabilityScope: observabilityScope
+        )
         artifactsToAdd.append(contentsOf: extractedLocalArtifacts)
 
         // Add the new artifacts
@@ -2411,333 +2425,10 @@ extension Workspace {
         observabilityScope.trap {
             try self.state.save()
         }
-    }
 
-    private func parseArtifacts(from manifests: DependencyManifests, observabilityScope: ObservabilityScope) throws -> (local: [ManagedArtifact], remote: [RemoteArtifact]) {
-        let packageAndManifests: [(reference: PackageReference, manifest: Manifest)] =
-            manifests.root.packages.values + // Root package and manifests.
-            manifests.dependencies.map({ manifest, managed, _, _ in (managed.packageRef, manifest) }) // Dependency package and manifests.
-
-        var localArtifacts: [ManagedArtifact] = []
-        var remoteArtifacts: [RemoteArtifact] = []
-
-        for (packageReference, manifest) in packageAndManifests {
-            for target in manifest.targets where target.type == .binary {
-                if let path = target.path {
-                    // TODO: find a better way to get the base path (not via the manifest)
-                    let absolutePath = try manifest.path.parentDirectory.appending(RelativePath(validating: path))
-                    localArtifacts.append(
-                        .local(
-                            packageRef: packageReference,
-                            targetName: target.name,
-                            path: absolutePath)
-                    )
-                } else if let url = target.url.flatMap(URL.init(string:)), let checksum = target.checksum {
-                    remoteArtifacts.append(
-                        .init(
-                            packageRef: packageReference,
-                            targetName: target.name,
-                            url: url,
-                            checksum: checksum)
-                    )
-                } else {
-                    throw InternalError("a binary target should have either a path or a URL and a checksum")
-                }
-            }
+        func isAtArtifactsDirectory(_ artifact: ManagedArtifact) -> Bool {
+            artifact.path.isDescendant(of: self.location.artifactsDirectory)
         }
-
-        return (local: localArtifacts, remote: remoteArtifacts)
-    }
-
-    private func download(_ artifacts: [RemoteArtifact], observabilityScope: ObservabilityScope) throws -> [ManagedArtifact] {
-        let group = DispatchGroup()
-        let result = ThreadSafeArrayStore<ManagedArtifact>()
-
-        // zip files to download
-        // stored in a thread-safe way as we may fetch more from "artifactbundleindex" files
-        let zipArtifacts = ThreadSafeArrayStore<RemoteArtifact>(artifacts.filter {
-            self.manifestLoader.supportedArchiveExtensions.contains($0.url.pathExtension.lowercased())
-        })
-
-        // fetch and parse "artifactbundleindex" files, if any
-        let indexFiles = artifacts.filter { $0.url.pathExtension.lowercased() == "artifactbundleindex" }
-        if !indexFiles.isEmpty {
-            let errors = ThreadSafeArrayStore<Error>()
-            let jsonDecoder = JSONDecoder.makeWithDefaults()
-            for indexFile in indexFiles {
-                group.enter()
-                var request = HTTPClient.Request(method: .get, url: indexFile.url)
-                request.options.validResponseCodes = [200]
-                request.options.authorizationProvider = self.authorizationProvider?.httpAuthorizationHeader(for:)
-                self.httpClient.execute(request) { result in
-                    defer { group.leave() }
-
-                    do {
-                        switch result {
-                        case .failure(let error):
-                            throw error
-                        case .success(let response):
-                            guard let body = response.body else {
-                                throw StringError("Body is empty")
-                            }
-                            // FIXME: would be nice if checksumAlgorithm.hash took Data directly
-                            let bodyChecksum = self.checksumAlgorithm.hash(ByteString(body)).hexadecimalRepresentation
-                            guard bodyChecksum == indexFile.checksum else {
-                                throw StringError("checksum of downloaded artifact of binary target '\(indexFile.targetName)' (\(bodyChecksum)) does not match checksum specified by the manifest (\(indexFile.checksum ))")
-                            }
-                            let metadata = try jsonDecoder.decode(ArchiveIndexFile.self, from: body)
-                            // FIXME: this filter needs to become more sophisticated
-                            guard let supportedArchive = metadata.archives.first(where: { $0.fileName.lowercased().hasSuffix(".zip") && $0.supportedTriples.contains(self.hostToolchain.triple) }) else {
-                                throw StringError("No supported archive was found for '\(self.hostToolchain.triple.tripleString)'")
-                            }
-                            // add relevant archive
-                            zipArtifacts.append(
-                                RemoteArtifact(
-                                    packageRef: indexFile.packageRef,
-                                    targetName: indexFile.targetName,
-                                    url: indexFile.url.deletingLastPathComponent().appendingPathComponent(supportedArchive.fileName),
-                                    checksum: supportedArchive.checksum)
-                            )
-                        }
-                    } catch {
-                        errors.append(error)
-                        observabilityScope.emit(error: "failed retrieving '\(indexFile.url)': \(error)")
-                    }
-                }
-            }
-
-            // wait for all "artifactbundleindex" files to be processed
-            group.wait()
-
-            // no reason to continue if we already ran into issues
-            if !errors.isEmpty {
-                throw Diagnostics.fatalError
-            }
-        }
-
-        // finally download zip files, if any
-        for artifact in zipArtifacts.get() {
-            let parentDirectory =  self.location.artifactsDirectory.appending(component: artifact.packageRef.identity.description)
-            guard observabilityScope.trap ({ try fileSystem.createDirectory(parentDirectory, recursive: true) }) else {
-                continue
-            }
-
-            let archivePath = parentDirectory.appending(component: artifact.url.lastPathComponent)
-            if self.fileSystem.exists(archivePath) {
-                guard observabilityScope.trap ({ try self.fileSystem.removeFileTree(archivePath) }) else {
-                    continue
-                }
-            }
-
-            group.enter()
-            var headers = HTTPClientHeaders()
-            headers.add(name: "Accept", value: "application/octet-stream")
-            var request = HTTPClient.Request.download(url: artifact.url, headers: headers, fileSystem: self.fileSystem, destination: archivePath)
-            request.options.authorizationProvider = self.authorizationProvider?.httpAuthorizationHeader(for:)
-            request.options.retryStrategy = .exponentialBackoff(maxAttempts: 3, baseDelay: .milliseconds(50))
-            request.options.validResponseCodes = [200]
-
-            let downloadStart: DispatchTime = .now()
-            self.delegate?.willDownloadBinaryArtifact(from: artifact.url.absoluteString)
-            observabilityScope.emit(debug: "downloading \(artifact.url) to \(archivePath)")
-            self.httpClient.execute(
-                request,
-                progress: { bytesDownloaded, totalBytesToDownload in
-                    self.delegate?.downloadingBinaryArtifact(
-                        from: artifact.url.absoluteString,
-                        bytesDownloaded: bytesDownloaded,
-                        totalBytesToDownload: totalBytesToDownload)
-                },
-                completion: { downloadResult in
-                    defer { group.leave() }
-
-                    // TODO: Use the same extraction logic for both remote and local archived artifacts.
-                    switch downloadResult {
-                    case .success:
-
-                        group.enter()
-                        observabilityScope.emit(debug: "validating \(archivePath)")
-                        self.archiver.validate(path: archivePath, completion: { validationResult in
-                            defer { group.leave() }
-
-                            switch validationResult {
-                            case .success(let valid):
-                                guard valid else {
-                                    observabilityScope.emit(.artifactInvalidArchive(artifactURL: artifact.url, targetName: artifact.targetName))
-                                    return
-                                }
-
-                                guard let archiveChecksum = observabilityScope.trap ({ try self.checksum(forBinaryArtifactAt: archivePath) }) else {
-                                    return
-                                }
-                                guard archiveChecksum == artifact.checksum else {
-                                    observabilityScope.emit(.artifactInvalidChecksum(targetName: artifact.targetName, expectedChecksum: artifact.checksum, actualChecksum: archiveChecksum))
-                                    observabilityScope.trap { try self.fileSystem.removeFileTree(archivePath) }
-                                    return
-                                }
-
-                                guard let tempExtractionDirectory = observabilityScope.trap({ () -> AbsolutePath in
-                                    let path = self.location.artifactsDirectory.appending(components: "extract", artifact.packageRef.identity.description, artifact.targetName, UUID().uuidString)
-                                    try self.fileSystem.forceCreateDirectory(at: path)
-                                    return path
-                                }) else {
-                                    return
-                                }
-
-                                group.enter()
-                                observabilityScope.emit(debug: "extracting \(archivePath) to \(tempExtractionDirectory)")
-                                self.archiver.extract(from: archivePath, to: tempExtractionDirectory, completion: { extractResult in
-                                    defer { group.leave() }
-
-                                    switch extractResult {
-                                    case .success:
-                                        var artifactPath: AbsolutePath? = nil
-                                        observabilityScope.trap {
-                                            try self.fileSystem.withLock(on: parentDirectory, type: .exclusive) {
-                                                // strip first level component if needed
-                                                if try self.fileSystem.shouldStripFirstLevel(archiveDirectory: tempExtractionDirectory, acceptableExtensions: BinaryTarget.Kind.allCases.map({ $0.fileExtension })) {
-                                                    observabilityScope.emit(debug: "stripping first level component from  \(tempExtractionDirectory)")
-                                                    try self.fileSystem.stripFirstLevel(of: tempExtractionDirectory)
-                                                } else {
-                                                    observabilityScope.emit(debug: "no first level component stripping needed for \(tempExtractionDirectory)")
-                                                }
-                                                let content = try self.fileSystem.getDirectoryContents(tempExtractionDirectory)
-                                                // copy from temp location to actual location
-                                                for file in content {
-                                                    let source = tempExtractionDirectory.appending(component: file)
-                                                    let destination = parentDirectory.appending(component: file)
-                                                    if self.fileSystem.exists(destination) {
-                                                        try self.fileSystem.removeFileTree(destination)
-                                                    }
-                                                    try self.fileSystem.copy(from: source, to: destination)
-                                                    if destination.basenameWithoutExt == artifact.targetName {
-                                                        artifactPath = destination
-                                                    }
-                                                }
-                                            }
-                                            // remove temp location
-                                            try self.fileSystem.removeFileTree(tempExtractionDirectory)
-                                        }
-
-                                        guard let mainArtifactPath = artifactPath else {
-                                            return observabilityScope.emit(.artifactNotFound(targetName: artifact.targetName, artifactName: artifact.targetName))
-                                        }
-
-                                        result.append(
-                                            .remote(
-                                                packageRef: artifact.packageRef,
-                                                targetName: artifact.targetName,
-                                                url: artifact.url.absoluteString,
-                                                checksum: artifact.checksum,
-                                                path: mainArtifactPath
-                                            )
-                                        )
-                                        self.delegate?.didDownloadBinaryArtifact(from: artifact.url.absoluteString, result: .success(mainArtifactPath), duration: downloadStart.distance(to: .now()))
-                                    case .failure(let error):
-                                        let reason = (error as? LocalizedError)?.errorDescription ?? "\(error)"
-                                        observabilityScope.emit(.artifactFailedExtraction(artifactURL: artifact.url, targetName: artifact.targetName, reason: reason))
-                                        self.delegate?.didDownloadBinaryArtifact(from: artifact.url.absoluteString, result: .failure(error), duration: downloadStart.distance(to: .now()))
-                                    }
-
-                                    observabilityScope.trap { try self.fileSystem.removeFileTree(archivePath) }
-                                })
-                            case .failure(let error):
-                                let reason = (error as? LocalizedError)?.errorDescription ?? "\(error)"
-                                observabilityScope.emit(.artifactFailedValidation(artifactURL: artifact.url, targetName: artifact.targetName, reason: "\(reason)"))
-                                self.delegate?.didDownloadBinaryArtifact(from: artifact.url.absoluteString, result: .failure(error), duration: downloadStart.distance(to: .now()))
-                            }
-                        })
-                    case .failure(let error):
-                        let reason = (error as? LocalizedError)?.errorDescription ?? "\(error)"
-                        observabilityScope.emit(.artifactFailedDownload(artifactURL: artifact.url, targetName: artifact.targetName, reason: "\(reason)"))
-                        self.delegate?.didDownloadBinaryArtifact(from: artifact.url.absoluteString, result: .failure(error), duration: downloadStart.distance(to: .now()))
-                    }
-                })
-        }
-
-        group.wait()
-
-        if zipArtifacts.count > 0 {
-            delegate?.didDownloadAllBinaryArtifacts()
-        }
-
-        return result.get()
-    }
-
-    private func extract(_ artifacts: [ManagedArtifact], observabilityScope: ObservabilityScope) throws -> [ManagedArtifact] {
-        let result = ThreadSafeArrayStore<ManagedArtifact>()
-        let group = DispatchGroup()
-
-        for artifact in artifacts {
-            let destinationDirectory = self.location.artifactsDirectory.appending(component: artifact.packageRef.identity.description)
-            try fileSystem.createDirectory(destinationDirectory, recursive: true)
-
-            let tempExtractionDirectory = self.location.artifactsDirectory.appending(components: "extract", artifact.packageRef.identity.description, artifact.targetName, UUID().uuidString)
-            try self.fileSystem.forceCreateDirectory(at: tempExtractionDirectory)
-
-            group.enter()
-            self.archiver.extract(from: artifact.path, to: tempExtractionDirectory, completion: { extractResult in
-                defer { group.leave() }
-
-                switch extractResult {
-                case .success:
-                    observabilityScope.trap { () -> Void in
-                        var artifactPath: AbsolutePath? = nil
-                        // strip first level component if needed
-                        if try self.fileSystem.shouldStripFirstLevel(archiveDirectory: tempExtractionDirectory, acceptableExtensions: BinaryTarget.Kind.allCases.map({ $0.fileExtension })) {
-                            observabilityScope.emit(debug: "stripping first level component from  \(tempExtractionDirectory)")
-                            try self.fileSystem.stripFirstLevel(of: tempExtractionDirectory)
-                        } else {
-                            observabilityScope.emit(debug: "no first level component stripping needed for \(tempExtractionDirectory)")
-                        }
-                        let content = try self.fileSystem.getDirectoryContents(tempExtractionDirectory)
-                        // copy from temp location to actual location
-                        for file in content {
-                            let source = tempExtractionDirectory.appending(component: file)
-                            let destination = destinationDirectory.appending(component: file)
-                            if self.fileSystem.exists(destination) {
-                                try self.fileSystem.removeFileTree(destination)
-                            }
-                            try self.fileSystem.copy(from: source, to: destination)
-                            if destination.basenameWithoutExt == artifact.targetName {
-                                artifactPath = destination
-                            }
-                        }
-
-                        // remove temp location
-                        try self.fileSystem.removeFileTree(tempExtractionDirectory)
-
-                        guard let mainArtifactPath = artifactPath else {
-                            return observabilityScope.emit(.localArchivedArtifactNotFound(targetName: artifact.targetName, artifactName: artifact.targetName))
-                        }
-
-                        // compute the checksum
-                        let artifactChecksum = try self.checksum(forBinaryArtifactAt: artifact.path)
-
-                        result.append(
-                            .local(
-                                packageRef: artifact.packageRef,
-                                targetName: artifact.targetName,
-                                path: mainArtifactPath,
-                                checksum: artifactChecksum
-                            )
-                        )
-                    }
-                case .failure(let error):
-                    let reason = (error as? LocalizedError)?.errorDescription ?? error.localizedDescription
-
-                    observabilityScope.emit(.localArtifactFailedExtraction(artifactPath: artifact.path, targetName: artifact.targetName, reason: reason))
-                }
-            })
-        }
-
-        group.wait()
-
-        return result.get()
-    }
-
-    private func isAtArtifactsDirectory(_ artifact: ManagedArtifact) -> Bool {
-        artifact.path.isDescendant(of: self.location.artifactsDirectory)
     }
 }
 
@@ -3452,37 +3143,6 @@ public final class LoadableResult<Value> {
     /// Load and return the value.
     public func load() throws -> Value {
         return try loadResult().get()
-    }
-}
-
-private struct RemoteArtifact {
-    let packageRef: PackageReference
-    let targetName: String
-    let url: URL
-    let checksum: String
-}
-
-private struct ArchiveIndexFile: Decodable {
-    let schemaVersion: String
-    let archives: [Archive]
-
-    struct Archive: Decodable {
-        let fileName: String
-        let checksum: String
-        let supportedTriples: [Triple]
-
-        enum CodingKeys: String, CodingKey {
-            case fileName
-            case checksum
-            case supportedTriples
-        }
-
-        public init(from decoder: Decoder) throws {
-            let container = try decoder.container(keyedBy: CodingKeys.self)
-            self.fileName = try container.decode(String.self, forKey: .fileName)
-            self.checksum = try container.decode(String.self, forKey: .checksum)
-            self.supportedTriples = try container.decode([String].self, forKey: .supportedTriples).map(Triple.init)
-        }
     }
 }
